feat: allow replacing all instances in a dir
run `stew replace <old_string> <new_string> [flags]` to replace all
instances of the old string with the new string in a project

if you want to ignore case, run
`stew replace <old_string> <new_string> [flags]` with the `-i` or
`--ignore-case` flag.

if you want to replace in only a specific directory,
specify the path with the `-p` or `--path` flag.